requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
decorator==5.1.1
diffusers==0.32.2
filelock==3.17.0
fsspec==2025.2.0
huggingface-hub==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.6.1
Jinja2==3.1.5
joblib==1.4.2
MarkupSafe==3.0.2
moviepy==2.1.2
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.2
opencv-python==4.11.0.86
packaging==24.2
pillow==10.3.0
proglog==0.1.10
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
setuptools==75.8.0
sympy==1.13.1
tokenizers==0.21.0
torch==2.6.0+cu126
torchaudio==2.6.0+cu126
torchvision==0.21.0+cu126
tqdm==4.67.1
transformers==4.48.3
triton-pre-nightly
typing_extensions==4.12.2
urllib3==2.3.0
webvtt-py==0.5.1
yt-dlp==2025.1.26
zipp==3.21.0
